Add deletion by look command

+++ src/main/java/net/ecoporium/ecoporium/command/EcoporiumCommand.java
@@ -5,6 +5,7 @@ import co.aikar.commands.BukkitCommandCompletionContext;
 import co.aikar.commands.CommandCompletions;
 import co.aikar.commands.CommandHelp;
 import co.aikar.commands.InvalidCommandArgument;
+import co.aikar.commands.annotation.CatchUnknown;
 import co.aikar.commands.annotation.CommandAlias;
 import co.aikar.commands.annotation.CommandPermission;
 import co.aikar.commands.annotation.Default;
@@ -16,9 +17,16 @@ import net.ecoporium.ecoporium.api.message.Message;
 import net.ecoporium.ecoporium.api.wrapper.Pair;
 import net.ecoporium.ecoporium.market.Market;
 import net.ecoporium.ecoporium.screen.TickerScreen;
+import org.bukkit.Bukkit;
+import org.bukkit.Material;
+import org.bukkit.block.Block;
 import org.bukkit.command.CommandSender;
+import org.bukkit.entity.EntityType;
+import org.bukkit.entity.ItemFrame;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.MapMeta;
+import org.bukkit.map.MapView;
 
 import java.util.LinkedList;
 import java.util.UUID;
@@ -88,6 +96,62 @@ public class EcoporiumCommand extends AbstractEcoporiumCommand {
 
         @Default
         public void onDelete(Player player, @Single TickerScreen tickerScreen) {
+            // remove from storage
+            plugin.getStorage().deleteTickerScreen(tickerScreen);
+            // remove from manager
+            plugin.getTickerScreenManager().removeScreen(tickerScreen.getId());
+
+            plugin.getMessages().ecoporiumDeleteDeleted.message(player);
+        }
+
+        @Subcommand("look")
+        public void onDeleteNoArgs(Player player) {
+            // get block they're looking at
+            Block targetBlock = player.getTargetBlockExact(20);
+
+            // if null
+            if (targetBlock == null) {
+                plugin.getMessages().ecoporiumDeleteCantFind.message(player);
+                return;
+            }
+
+            // get near entities / item frames
+            ItemFrame itemFrame = targetBlock.getWorld().getNearbyEntities(targetBlock.getLocation(), 1d, 1d, 1d).stream()
+                    .filter(e -> e instanceof ItemFrame)
+                    .map(e -> (ItemFrame) e)
+                    .findFirst()
+                    .orElse(null);
+
+            if (itemFrame == null) {
+                plugin.getMessages().ecoporiumDeleteCantFind.message(player);
+                return;
+            }
+
+            ItemStack item = itemFrame.getItem();
+
+            if (!item.getType().equals(Material.FILLED_MAP)) {
+                plugin.getMessages().ecoporiumDeleteCantFind.message(player);
+                return;
+            }
+
+            // get map id
+            MapMeta meta = (MapMeta) item.getItemMeta();
+
+            if (meta == null) {
+                plugin.getMessages().ecoporiumDeleteCantFind.message(player);
+                return;
+            }
+
+            int mapId = meta.getMapView().getId();
+
+            // find screen by mapId
+            TickerScreen tickerScreen = plugin.getTickerScreenManager().findByMapId(mapId);
+
+            if (tickerScreen == null) {
+                plugin.getMessages().ecoporiumDeleteCantFind.message(player);
+                return;
+            }
+
             // remove from storage
             plugin.getStorage().deleteTickerScreen(tickerScreen);
             // remove from manager
